refactor(fetchwrapper): add typs in login()
