+++ jest.config.js
@@ -3,4 +3,6 @@ module.exports = {
    preset: "ts-jest",
    testEnvironment: "node",
    setupFiles: ["dotenv/config"],
+   moduleFileExtensions: ["ts", "js"],
+   moduleDirectories: ["node_modules", "src"],
 }
